Make more steps separate in input screen
