+++ src/pages/Chat/state.tsx
@@ -4,11 +4,13 @@ import { notNullable } from '../../utils/nullables'
 import { StorageContainer } from '../../utils/StorageContainer'
 
 export enum ChatStep {
+  NUMBER_OF_PEOPLE = 'NUMBER_OF_PEOPLE',
   INITIAL = 'INITIAL',
   NAME_INPUT = 'NAME_INPUT',
   AGE_INPUT = 'AGE_INPUT',
   GREET = 'GREET',
-  LIVING_SITUATION_INPUT = 'LIVING_SITUATION_INPUT',
+  ADDRESS_INPUT = 'ADDRESS_INPUT',
+  APARTMENT_TYPE_INPUT = 'APARTMENT_TYPE_INPUT',
   CURRENT_INSURANCE_QUESTION = 'CURRENT_INSURANCE_QUESTION',
   CURRENT_INSURANCE_INPUT = 'CURRENT_INSURANCE_INPUT',
   SHOW_OFFER = 'SHOW_OFFER',
