+++ src/pages/Chat/steps/AddressInput.test.tsx
@@ -5,8 +5,7 @@ import { MockTextKeyProvider } from 'utils/MockTextKeyProvider'
 import { createSession } from 'utils/sessionStorage'
 import { MockStorage } from 'utils/storage/MockStorage'
 import { WithStorageProps } from 'utils/StorageContainer'
-import { Insurer } from '../state'
-import { CurrentInsuranceInput } from './CurrentInsuranceInput'
+import { AddressInput } from './AddressInput'
 
 it('handles form changes', () => {
   const wrapper = mount(
@@ -15,12 +14,10 @@ it('handles form changes', () => {
     >
       <MockTextKeyProvider
         textKeys={{
-          CHAT_INPUT_CURRENT_INSURANCE_TEXT: '{toggle}',
-          CHAT_INPUT_CURRENT_INSURANCE_TEXT_HAS_INSURANCE:
-            '{toggle} {currentInsurerMaybe}',
+          CHAT_INPUT_ADDRESS_TEXT: '{streetAddress} {postalCode}',
         }}
       >
-        <CurrentInsuranceInput isCurrentMessage />
+        <AddressInput isCurrentMessage />
       </MockTextKeyProvider>
     </Provider>,
   )
@@ -28,14 +25,13 @@ it('handles form changes', () => {
   expect(wrapper.find('button')).toHaveLength(0)
 
   wrapper
-    .find('select#hasCurrentInsurance')
-    .simulate('change', { target: { value: 'yes' } })
-  expect(wrapper.find('select#hasCurrentInsurance').prop('value')).toBe('yes')
+    .find('input#streetAddress')
+    .simulate('change', { target: { value: 'Storgatan 1' } })
+  expect(wrapper.find('input#streetAddress').prop('value')).toBe('Storgatan 1')
   wrapper
-    .find('select#currentInsurer')
-    .simulate('change', { target: { value: Insurer.FOLKSAM } })
-  expect(wrapper.find('select#currentInsurer').prop('value')).toBe(
-    Insurer.FOLKSAM,
-  )
+    .find('input#postalCode')
+    .simulate('change', { target: { value: '12345' } })
+  expect(wrapper.find('input#postalCode').prop('value')).toBe('12345')
+
   expect(wrapper.find('button')).toHaveLength(1)
 })
