+++ src/pages/Chat/steps/ApartmentTypeInput.test.tsx
@@ -7,7 +7,7 @@ import { createSession } from 'utils/sessionStorage'
 import { MockStorage } from 'utils/storage/MockStorage'
 import { WithStorageProps } from 'utils/StorageContainer'
 import { ApartmentType } from '../state'
-import { LivingSituationInput } from './LivingSituationInput'
+import { ApartmentTypeInput } from './ApartmentTypeInput'
 
 it('handles form changes', () => {
   const wrapper = mount(
@@ -16,34 +16,26 @@ it('handles form changes', () => {
     >
       <MockTextKeyProvider
         textKeys={{
-          CHAT_INPUT_LIVING_SITUATION_TEXT:
-            '{streetAddress} {postalCode} {apartmentType} {size} {sizeValidationErrorMaybe} {numberOfPeople}',
+          CHAT_INPUT_APARTMENT_TYPE_TEXT: '{apartmentType} {size}',
         }}
       >
-        <LivingSituationInput />
+        <ApartmentTypeInput isCurrentMessage />
       </MockTextKeyProvider>
     </Provider>,
   )
-  wrapper
-    .find('input#streetAddress')
-    .simulate('change', { target: { value: 'Storgatan 1' } })
-  expect(wrapper.find('input#streetAddress').prop('value')).toBe('Storgatan 1')
-  wrapper
-    .find('input#postalCode')
-    .simulate('change', { target: { value: '12345' } })
-  expect(wrapper.find('input#postalCode').prop('value')).toBe('12345')
+
+  expect(wrapper.find('button')).toHaveLength(0)
+
   wrapper
     .find('select#apartmentType')
     .simulate('change', { target: { value: ApartmentType.RENT } })
   expect(wrapper.find('select#apartmentType').prop('value')).toBe(
     ApartmentType.RENT,
   )
-  wrapper.find('input#size').simulate('change', { target: { value: 54 } })
-  expect(wrapper.find('input#size').prop('value')).toBe(54)
-  wrapper
-    .find('select#numberOfPeople')
-    .simulate('change', { target: { value: 3 } })
-  expect(wrapper.find('select#numberOfPeople').prop('value')).toBe(3)
+  wrapper.find('input#size').simulate('change', { target: { value: 42 } })
+  expect(wrapper.find('input#size').prop('value')).toBe(42)
+
+  expect(wrapper.find('button')).toHaveLength(1)
 })
 
 it('handles too big apartment', () => {
@@ -53,11 +45,10 @@ it('handles too big apartment', () => {
     >
       <MockTextKeyProvider
         textKeys={{
-          CHAT_INPUT_LIVING_SITUATION_TEXT:
-            '{streetAddress} {postalCode} {apartmentType} {size} {sizeValidationErrorMaybe} {numberOfPeople}',
+          CHAT_INPUT_APARTMENT_TYPE_TEXT: '{apartmentType} {size}',
         }}
       >
-        <LivingSituationInput />
+        <ApartmentTypeInput />
       </MockTextKeyProvider>
     </Provider>,
   )
