+++ src/pages/Chat/steps/Greet.tsx
@@ -1,3 +1,4 @@
+import { TranslationsPlaceholderConsumer } from '@hedviginsurance/textkeyfy'
 import { ChatMessage, ChatMessageProps } from 'components/hedvig/chat'
 import * as React from 'react'
 import { ChatContainer } from '../state'
@@ -6,8 +7,14 @@ export const Greet: React.SFC<ChatMessageProps> = ({ onTyped, appear }) => (
   <ChatContainer>
     {({ nameAge }) => (
       <ChatMessage appear={appear} typingDuration={500} onTyped={onTyped}>
-        Kul att ha dig här {(nameAge && nameAge.firstName) || ''}!<br />
-        Hur bor du?
+        <TranslationsPlaceholderConsumer
+          textKey="CHAT_HEDVIG_GREET"
+          replacements={{
+            name: nameAge.firstName,
+          }}
+        >
+          {(text) => text}
+        </TranslationsPlaceholderConsumer>
       </ChatMessage>
     )}
   </ChatContainer>
